feat(project): typescript
All components now have a proper React.FC type. The contexts have been rewritten in typescript. An @ alias has been setup for absolute paths.
