--- sccprogrammingclub.com-frontend/src/components/Loader/Loader.jsx
@@ -1,11 +1,0 @@
-import "./Loader.css";
-
-function Loader() {
-  return (
-    <div id="loader-overlay">
-      <div className="spinner"></div>
-    </div>
-  );
-}
-
-export default Loader;
