+++ sccprogrammingclub.com-frontend/src/components/Loader/Loader.tsx
@@ -1,0 +1,9 @@
+import "./Loader.css";
+
+export const Loader = () => {
+    return (
+        <div id="loader-overlay">
+            <div className="spinner" />
+        </div>
+    );
+}
