+++ sccprogrammingclub.com-frontend/src/components/shared/Footer/Footer.tsx
@@ -1,0 +1,23 @@
+import "./Footer.css";
+
+export const Footer: React.FC = () => {
+    return (
+        <footer>
+        <p className="footer-text">Programming Club (2022 - 2025)</p>
+        <div className="footer-icons">
+            <a
+            href="https://github.com/sccprogrammingclub"
+            className="icon-link github"
+            >
+            <img src="src/assets/icons/github.svg" alt="GitHub" />
+            </a>
+            <a
+            href="https://www.instagram.com/programmingclub.scc"
+            className="icon-link instagram"
+            >
+            <img src="src/assets/icons/instagram.svg" alt="Instagram" />
+            </a>
+        </div>
+        </footer>
+    );
+}
