--- sccprogrammingclub.com-frontend/src/contexts/auth/AuthContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const AuthContext = createContext();
-
-export default AuthContext;
