--- sccprogrammingclub.com-frontend/src/contexts/auth/auth.js
@@ -1,8 +1,0 @@
-import AuthContext from "./AuthContext";
-import { useContext } from "react";
-
-function useAuth() {
-  return useContext(AuthContext);
-}
-
-export default useAuth;
