--- sccprogrammingclub.com-frontend/src/contexts/member/MemberContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const MemberContext = createContext();
-
-export default MemberContext;
