--- sccprogrammingclub.com-frontend/src/contexts/member/member.js
@@ -1,8 +1,0 @@
-import { useContext } from "react";
-import MemberContext from "./MemberContext";
-
-function useMember() {
-    return useContext(MemberContext);
-}
-
-export default useMember;
