+++ sccprogrammingclub.com-frontend/src/main.tsx
@@ -4,7 +4,9 @@ import './index.css'
 import { BrowserRouter } from 'react-router-dom'
 import App from './App.jsx'
 
-createRoot(document.getElementById('root')).render(
+const root = document.getElementById('root')!;
+
+createRoot(root).render(
   <StrictMode>
     <BrowserRouter>
       <App />
