--- sccprogrammingclub.com-frontend/src/pages/Projects/Projects.jsx
@@ -1,25 +1,0 @@
-import "./Projects.css";
-
-function Projects() {
-  return (
-    <div className="projects-page">
-      <h2>Here is a list of the projects that we have worked on:</h2>
-      <ul>
-        <li>
-          <a href="">Powdertoy clone</a>
-        </li>
-        <li>
-          <a href="">Snake game</a>
-        </li>
-        <li>
-          <a href="">Google</a>
-        </li>
-        <li>
-          <a href="">Sample</a>
-        </li>
-      </ul>
-    </div>
-  );
-}
-
-export default Projects;
