+++ sccprogrammingclub.com-frontend/src/pages/Projects/Projects.tsx
@@ -1,0 +1,23 @@
+import "./Projects.css";
+
+export const Projects: React.FC = () => {
+    return (
+        <div className="projects-page">
+        <h2>Here is a list of the projects that we have worked on:</h2>
+        <ul>
+            <li>
+            <a href="">Powdertoy clone</a>
+            </li>
+            <li>
+            <a href="">Snake game</a>
+            </li>
+            <li>
+            <a href="">Google</a>
+            </li>
+            <li>
+            <a href="">Sample</a>
+            </li>
+        </ul>
+        </div>
+    );
+}
